requirements.txt
streamlit>=1.32.0
google-genai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pillow>=10.0.0
reportlab>=4.0.0  # For PDF export in app.py